% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcparallelDo.R
\name{checkIfJobStillRunning}
\alias{checkIfJobStillRunning}
\title{checkIfJobStillRunning}
\usage{
checkIfJobStillRunning(targetJob, targetValue, verbose, targetEnvironment)
}
\arguments{
\item{targetJob}{(character) The job name}

\item{targetValue}{(character) The return variable name}

\item{verbose}{(logical) Whether a message will be generated when complete}

\item{targetEnvironment}{(environment) Target environment}
}
\value{
logical; TRUE if still running; FALSE if not running
}
\description{
checkIfJobStillRunning
}

