% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcparallelDo.R
\name{mcparallelDoCheck}
\alias{mcparallelDoCheck}
\title{mcparallelDoCheck}
\usage{
mcparallelDoCheck()
}
\value{
A named logical vector, TRUE if complete, FALSE if not complete
}
\description{
Forces a check on all mcparallelDo jobs and returns their values to the target environment if they are complete.
}

