% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcparallelDo.R
\docType{class}
\name{mcparallelDoManagerClass}
\alias{mcparallelDoManager}
\alias{mcparallelDoManagerClass}
\title{The mcparallelDoManager Class and Object}
\format{\preformatted{Class 'R6ClassGenerator' <mcparallelDoManager> object generator
  Public:
    h: list
    runningJobs: list
    addJob: function (jobName, targetValue, verbose, targetEnvironment) 
    removeJob: function (x) 
    checkJobs: function () 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:mcparallelDo>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "mcparallelDoManager_generator"
}}
\usage{
mcparallelDoManagerClass
}
\description{
The mcparallelDoManager Class and Object
}
\keyword{datasets}

