% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcparallelDo.R
\docType{package}
\name{mcparallelDo-package}
\alias{mcparallelDo-package}
\title{mcparallelDo-package placeholder}
\description{
Asynchronous Exploritory Data Analysis
}
\details{
The primary function of this package is mcparallelDo().  
To use mcparallelDo(), simply invoke the function with a curly braced wrapped code and the character element name to which you want to assign the results.
}

