% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcparallelDo.R
\docType{class}
\name{mcparallelDoManagerClass}
\alias{mcparallelDoManager}
\alias{mcparallelDoManagerClass}
\title{The mcparallelDoManager Class and Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
mcparallelDoManagerClass
}
\description{
The mcparallelDoManager Class and Object
}
\keyword{datasets}

